Reagents ui in typescript (#89088) This is the reagent library that can be accessed from the dispenser. Cool ui btw!
Removes uselocalstate as its deprecated
Converts the entire UI to typescript
Breaks up the file into a directory (its >500 LOC)
Some code stylistic choices ie arrow functions -> function declarations
Code quality + typescript
N/A
